+++ changelog.md
@@ -1,3 +1,16 @@
 # Changelog
 
-All output successful
+**./src/common/csv-errors.js**
+* Simplified error text writing.
+
+---
+
+**./src/common/fs-errors.js**
+* Simplified error text writing for these functions:
+	* writeFileActionErrorText
+	* writeFolderActionErrorText
+
+---
+
+**./src/output/save-json.js - writeStringifyErrorText**
+* Simplified error text writing.